removed modules Authentication, Strategies
