--- lib/jump_in/strategies.rb
@@ -1,3 +1,0 @@
-require 'jump_in/strategies/by_password'
-require 'jump_in/strategies/session'
-require 'jump_in/strategies/cookies'
